packageが異なると、structの構成が同じでもtype assertionでエラーになってしまうので、modelに移動
